refactor(rest): extract streaming csv/xlsx tests to sibling tests.rs (#269)
CI's workspace-wide `lint-tests-layout` rule (Makefile:412) rejects inline
`mod tests {}` blocks anywhere under `crates/*/src/` except files literally
named `tests.rs`. The CSV cycles introduced one inline test block in
`streaming/csv.rs`, and Cycle 6 added another in `streaming/xlsx.rs`. Both
slipped past local verification because the rule lives in `make
lint-tests-layout` rather than in `cargo` — the CSV cycles never ran as a
standalone PR, so this is the first CI run that caught either.

The fix converts each module from a flat file into a directory module that
follows the NDJSON sibling's convention (`streaming/mod.rs` +
`streaming/tests.rs`):

    streaming/csv.rs   →  streaming/csv/mod.rs   + streaming/csv/tests.rs
    streaming/xlsx.rs  →  streaming/xlsx/mod.rs  + streaming/xlsx/tests.rs

No test logic changes. Tests still access private items in their parent
module via `use super::*;` exactly as the inline blocks did — a child
`mod tests;` declared from `mod.rs` has the same privacy access as an
inline `mod tests {}`. File-level `#![allow(clippy::unwrap_used)]`
replaces the per-block `#[allow]` to match `streaming/tests.rs`.

## Changes
- New: `streaming/csv/mod.rs` (impl, was csv.rs lines 1-498) +
  `streaming/csv/tests.rs` (51 tests, was csv.rs lines 503-993).
- New: `streaming/xlsx/mod.rs` (impl, was xlsx.rs lines 1-494) +
  `streaming/xlsx/tests.rs` (38 tests, was xlsx.rs lines 499-842).
- Each new `mod.rs` declares `#[cfg(test)] mod tests;` (declaration
  form, no inline `{` — passes the layout rule).
- Old `streaming/csv.rs` and `streaming/xlsx.rs` deleted.

## Verification
- make lint-tests-layout → "OK: no inline test blocks in workspace src/"
- cargo check -p fraiseql-server --features rest clean
- cargo check -p fraiseql-server --features rest,export-csv,export-xlsx clean
- cargo clippy -p fraiseql-server --features rest,export-csv,export-xlsx --all-targets -- -D warnings clean
- cargo nextest run -p fraiseql-server --features rest,export-csv,export-xlsx --lib → 1464/1464 passed
- cargo +nightly fmt -p fraiseql-server -- --check clean
